client: merge useUser and useMember hooks
